Testing out database and ensuring GET requests work for retrieving books and authors. Service, Control and API layers added

+++ src/main/java/com/example/demo/library/Borrow.java
@@ -4,8 +4,8 @@ import java.time.LocalDate;
 
 public class Borrow extends Book{
 
-    private int id;
-    private int BookID;
+    private Long id;
+    private Long BookID;
     private String username;
     private LocalDate DateTime;
